requirements.txt
streamlit
pandas
matplotlib
Pillow